+++ test/sys_tests/test_sys_block_devices.py
@@ -1,6 +1,7 @@
 #!/usr/bin/python3
 
 import troposphere.elasticloadbalancing as elb
+from amazonia.classes.sns import SNS
 from amazonia.classes.asg import Asg
 from amazonia.classes.asg_config import AsgConfig
 from amazonia.classes.block_devices_config import BlockDevicesConfig
@@ -43,6 +44,8 @@ def main():
         def __init__(self):
             self.single = ec2.Instance('title')
 
+    sns_topic = SNS(template)
+
     network_config = NetworkConfig(
         vpc=vpc,
         private_subnets=subnets,
@@ -55,7 +58,8 @@ def main():
         keypair='pipeline',
         cd_service_role_arn='arn:aws:iam::12345678987654321:role/CodeDeployServiceRole',
         nat_highly_available=False,
-        nat_gateways=[]
+        nat_gateways=[],
+        sns_topic=sns_topic
     )
 
     block_devices_config = [
@@ -94,11 +98,6 @@ def main():
         health_check_grace_period=300,
         health_check_type='ELB',
         iam_instance_profile_arn='arn:aws:iam::12345678987654321:role/InstanceProfileRole',
-        sns_topic_arn='arn:aws:sns:ap-southeast-2:123456789:test_sns_arn',
-        sns_notification_types=['autoscaling:EC2_INSTANCE_LAUNCH',
-                                'autoscaling:EC2_INSTANCE_LAUNCH_ERROR',
-                                'autoscaling:EC2_INSTANCE_TERMINATE',
-                                'autoscaling:EC2_INSTANCE_TERMINATE_ERROR'],
         block_devices_config=block_devices_config,
         simple_scaling_policy_config=None
     )