+++ test/sys_tests/test_sys_elb.py
@@ -7,6 +7,7 @@ from amazonia.classes.hosted_zone import HostedZone
 from amazonia.classes.network_config import NetworkConfig
 from amazonia.classes.single_instance import SingleInstance
 from amazonia.classes.single_instance_config import SingleInstanceConfig
+from amazonia.classes.sns import SNS
 from troposphere import ec2, Ref, Tags, Template
 
 
@@ -48,6 +49,8 @@ def main():
                                                         AvailabilityZone='ap-southeast-2c',
                                                         VpcId=Ref(vpc),
                                                         CidrBlock='10.0.6.0/24'))]
+    sns_topic = SNS(template)
+
     single_instance_config = SingleInstanceConfig(
         keypair='pipeline',
         si_image_id='ami-53371f30',
@@ -55,11 +58,10 @@ def main():
         vpc=vpc,
         subnet=public_subnets[0],
         instance_dependencies=vpc.title,
-        alert=None,
-        alert_emails=None,
         public_hosted_zone_name=None,
         iam_instance_profile_arn=None,
-        is_nat=True
+        is_nat=True,
+        sns_topic=sns_topic
     )
     nat = SingleInstance(title='Nat',
                          template=template,
@@ -77,7 +79,8 @@ def main():
         keypair=None,
         cd_service_role_arn=None,
         nat_highly_available=False,
-        nat_gateways=[]
+        nat_gateways=[],
+        sns_topic=sns_topic
     )
 
     elb_listeners_config = [
@@ -105,7 +108,7 @@ def main():
         unhealthy_threshold=2,
         interval=300,
         timeout=30,
-        sticky_app_cookies=['JSESSION','SESSIONTOKEN']
+        sticky_app_cookies=['JSESSION', 'SESSIONTOKEN']
     )
     elb_config2 = ElbConfig(
         elb_listeners_config=elb_listeners_config,
@@ -117,7 +120,7 @@ def main():
         unhealthy_threshold=2,
         interval=300,
         timeout=30,
-        sticky_app_cookies=['JSESSION','SESSIONTOKEN']
+        sticky_app_cookies=['JSESSION', 'SESSIONTOKEN']
     )
     elb_config3 = ElbConfig(
         elb_listeners_config=elb_listeners_config,
@@ -129,7 +132,7 @@ def main():
         unhealthy_threshold=2,
         interval=300,
         timeout=30,
-        sticky_app_cookies=['JSESSION','SESSIONTOKEN']
+        sticky_app_cookies=['JSESSION', 'SESSIONTOKEN']
     )
 
     Elb(title='MyUnit1',