+++ test/sys_tests/test_sys_single_nat.py
@@ -1,9 +1,9 @@
 #!/usr/bin/env python
 
-from troposphere import ec2, Ref, Template
-
 from amazonia.classes.single_instance import SingleInstance
 from amazonia.classes.single_instance_config import SingleInstanceConfig
+from amazonia.classes.sns import SNS
+from troposphere import ec2, Ref, Template
 
 
 def main():
@@ -14,6 +14,7 @@ def main():
                         VpcId=Ref(vpc),
                         CidrBlock='10.0.1.0/24')
     template = Template()
+    sns_topic = SNS(template)
     single_instance_config = SingleInstanceConfig(
         keypair='pipeline',
         si_image_id='ami-53371f30',
@@ -23,9 +24,8 @@ def main():
         is_nat=True,
         instance_dependencies=vpc.title,
         public_hosted_zone_name=None,
-        alert=False,
-        alert_emails=[],
-        iam_instance_profile_arn=None
+        iam_instance_profile_arn=None,
+        sns_topic=sns_topic
     )
     SingleInstance(title='nat1',
                    template=template,